feat(nh62): CharacterClass structural refactor for multiclass (Phase 1)
CharacterClass changes shape from a tuple struct wrapping one ClassKind to
{ primary, secondary: Option<ClassKind> } + single()/classes()/has()/
is_multiclass() helpers. Deliberately a breaking-compilation change (not an
additive component) so the compiler enumerates every one of the ~11
affected call sites -- an additive SecondaryClass component would compile
everywhere and silently keep every existing gate single-class-only forever.

Every call site keeps single-class semantics in this PR (secondary is
always None) -- zero observable behavior change. Adds a nullable
character.secondary_class column (migration V75) and its persistence
converters, reusing the existing class_to_db_string/db_string_to_class
verbatim.

+++ server/src/character_creator.rs
@@ -165,7 +165,7 @@ pub fn create_character(
     character_updater.create_character(entity, player_uuid, character_alias, PersistedComponents {
         body,
         hardcore: hardcore.then_some(common::comp::Hardcore),
-        character_class: CharacterClass(character_class),
+        character_class: CharacterClass::single(character_class),
         stats,
         skill_set,
         inventory,
